Cleanup samples and add BotController and readme.md to samples
