+++ samples/AlarmBot-Cards/BotController.cs
@@ -6,13 +6,12 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Adapters;
-using Microsoft.Bot.Builder.Adapters;
 using Microsoft.Bot.Schema;
 
-namespace AlarmBot
+namespace Microsoft.Bot.Builder
 {
     /// <summary>
-    /// Helper controller for boiler plate adapter usage
+    /// Helper Bot Controller for ASP.NET Core
     /// </summary>
     public class BotController : Controller
     {
