+++ samples/Microsoft.Bot.Samples.CustomMiddleware/readme.md
@@ -1,0 +1,4 @@
+﻿# Custom Middleware sample
+This sample shows a Microsoft.Bot.Builder middleware class ( ~Middleware/ExampleMiddleware.cs ) which writes to log on every incoming and outgoing 
+activity
+
