+++ samples/Microsoft.Bot.Samples.EchoBot-AspNet461/Microsoft.Bot.Samples.EchoBot-AspNet461.csproj
@@ -147,6 +147,7 @@
     <Compile Include="App_Start\UnityConfig.cs" />
     <Compile Include="App_Start\UnityWebApiActivator.cs" />
     <Compile Include="App_Start\WebApiConfig.cs" />
+    <Compile Include="BotController.cs" />
     <Compile Include="Controllers\MessagesController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
@@ -157,6 +158,7 @@
     <Content Include="packages.config">
       <SubType>Designer</SubType>
     </Content>
+    <Content Include="readme.md" />
     <None Include="Web.Debug.config">
       <DependentUpon>Web.config</DependentUpon>
     </None>
@@ -168,6 +170,10 @@
     <ProjectReference Include="..\..\libraries\Microsoft.Bot.Builder.BotFramework\Microsoft.Bot.Builder.BotFramework.csproj">
       <Project>{20DB2748-6562-472F-AD20-E57167913C7D}</Project>
       <Name>Microsoft.Bot.Builder.BotFramework</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\..\libraries\Microsoft.Bot.Builder.Core\Microsoft.Bot.Builder.Core.csproj">
+      <Project>{f0c150e2-9b8a-444c-b70e-97ad918cb3d4}</Project>
+      <Name>Microsoft.Bot.Builder.Core</Name>
     </ProjectReference>
     <ProjectReference Include="..\..\libraries\Microsoft.Bot.Builder\Microsoft.Bot.Builder.csproj">
       <Project>{74908F42-3842-49B4-9344-60BB791FDE32}</Project>
