+++ samples/experiments/Microsoft.Bot.Samples.Simplified.Asp/MessageActivityAttribute.cs
